requirements.txt
numpy
pillow
pyowm
